Bug 574980 Passage interface implementations should also be published
publish XmiConditionTransport
