+++ bundles/org.eclipse.passage.lic.licenses.model/META-INF/MANIFEST.MF
@@ -14,17 +14,13 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0";x-installation:=
  org.eclipse.passage.lic.emf;bundle-version="2.0.0",
  org.eclipse.passage.lic.licenses;bundle-version="0.0.0";visibility:=reexport,
  org.eclipse.passage.lic.base;bundle-version="0.0.0"
-Export-Package: org.eclipse.passage.lic.internal.licenses.convert;
-  x-friends:="org.eclipse.passage.lic.hc",
- org.eclipse.passage.lic.internal.licenses.model;
-  x-friends:="org.eclipse.passage.lic.licenses.migration,
-   org.eclipse.passage.loc.dashboard.ui,
-   org.eclipse.passage.loc.licenses.core",
+Export-Package: org.eclipse.passage.lic.internal.licenses.convert;x-friends:="org.eclipse.passage.lic.hc",
+ org.eclipse.passage.lic.internal.licenses.model;x-friends:="org.eclipse.passage.lic.licenses.migration,org.eclipse.passage.loc.dashboard.ui,org.eclipse.passage.loc.licenses.core",
  org.eclipse.passage.lic.internal.licenses.model.signature,
- org.eclipse.passage.lic.internal.licenses.model.toberemoved;x-internal:=true,
  org.eclipse.passage.lic.licenses.model.api,
  org.eclipse.passage.lic.licenses.model.impl;x-internal:=true,
  org.eclipse.passage.lic.licenses.model.meta,
+ org.eclipse.passage.lic.licenses.model.transport,
  org.eclipse.passage.lic.licenses.model.util
 Bundle-ClassPath: .
 Bundle-ActivationPolicy: lazy
