+++ tests/org.eclipse.passage.lic.hc.tests/src/org/eclipse/passage/lic/internal/hc/tests/remote/TestEquipment.java
@@ -27,7 +27,7 @@ import org.eclipse.passage.lic.base.io.PathKeyKeeper;
 import org.eclipse.passage.lic.base.registry.ReadOnlyRegistry;
 import org.eclipse.passage.lic.bc.BcStreamCodec;
 import org.eclipse.passage.lic.hc.remote.impl.Equipment;
-import org.eclipse.passage.lic.internal.licenses.model.toberemoved.XmiConditionTransport;
+import org.eclipse.passage.lic.licenses.model.transport.XmiConditionTransport;
 
 @SuppressWarnings("restriction")
 final class TestEquipment implements Supplier<Equipment> {
