+++ tests/org.eclipse.passage.lic.licenses.model.tests/src/org/eclipse/passage/lic/internal/licenses/model/migration/tests/XmiLicensingConditionExtractorTest.java
@@ -27,10 +27,10 @@ import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.passage.lic.api.conditions.Condition;
 import org.eclipse.passage.lic.internal.licenses.model.EmptyPersonalFeatureGrant;
 import org.eclipse.passage.lic.internal.licenses.model.EmptyPersonalLicensePack;
-import org.eclipse.passage.lic.internal.licenses.model.toberemoved.XmiConditionTransport;
 import org.eclipse.passage.lic.licenses.model.api.PersonalFeatureGrant;
 import org.eclipse.passage.lic.licenses.model.api.PersonalLicensePack;
 import org.eclipse.passage.lic.licenses.model.api.ValidityPeriodClosed;
+import org.eclipse.passage.lic.licenses.model.transport.XmiConditionTransport;
 import org.eclipse.passage.lic.licenses.model.util.LicensesResourceImpl;
 import org.junit.Rule;
 import org.junit.Test;
